Create GUI + model.py
This file has the user interface + model architecture. my saved model was h5 not keras so when using the interface, you should apply your architecture. if you don't know the way you preprocessed or the exact way of architecture use the keras file for this part. this code will automatically handle both files.