Add files via upload
GUI look, I've added the part which shows the three features that are in the ensemble. this will be unnecessary for the train of model and, it's just used for my check. in the code of GUI I just load the Mel Spectrogram and then find the other features that I extracted at past by their names