Create every_backbone_model.py
this python file has every backbone used in the thesis, and got more of them. but some backbones which was deep or didn't get good accuracy was omitted there. this code will train 7 different structures of models and it will train them for all three features, at last for the spectral contrast we used a simple CNN2D that will be added to the GitHub. this code has the hyper tunes which are similar to the RenNet50 but we changes them a little bit based o their differences of the structure.